# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: WizLetter API
  description: |
    WizLetter (위즈레터) backend API — AI-curated daily briefing service with coin/level gamification.

    ## Authentication
    Most endpoints require a JWT access token stored in an HttpOnly cookie (`ota_token`).
    The token is issued automatically after Kakao OAuth login. A short-lived access token
    (15 min) and a long-lived refresh token (7 days, cookie `ota_refresh`) are issued together.

    Endpoints may also accept `Authorization: Bearer <token>` as a fallback.

    ## Base URL
    - Production: `https://server.mindhacker.club`
    - Local: `http://localhost:8080`
  version: "1.0.0"
  contact:
    email: mindhacker.admin@gmail.com
servers:
  - url: https://server.mindhacker.club/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ota_token
      description: JWT access token (15 min TTL). Set automatically after Kakao login.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Fallback bearer token (same JWT as cookie).
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        kakao_id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        nickname:
          type: string
        profile_image:
          type: string
        role:
          type: string
          enum: [user, admin]
        email_verified:
          type: boolean
...
paths:
...
  /auth/kakao/callback:
    get:
      tags: [auth]
      summary: Kakao OAuth callback
      description: |
        Handles the OAuth redirect from Kakao. For existing users, issues auth cookies
        and redirects to the frontend home page. For new users, caches a pending signup
        and redirects to `/terms-consent?signup_key=<uuid>`.
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect to frontend (home or terms-consent page)
  /auth/complete-signup:
    post:
      tags: [auth]
      summary: Complete two-phase signup with terms consent
      description: |
        Finalises new-user registration. The `signup_key` was issued by the Kakao callback
        and cached for 3 minutes. Required terms must be included in `agreed_term_ids`.
        On success, issues auth cookies and returns the new user object.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [signup_key, agreed_term_ids]
              properties:
                signup_key:
                  type: string
                  format: uuid
                  description: One-time key received from the Kakao callback redirect
                agreed_term_ids:
                  type: array
                  items:
                    type: string
                  description: IDs of terms the user has agreed to
      responses:
        '200':
          description: Signup complete; auth cookies set
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Missing/expired signup key, missing required terms, or invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags: [auth]
      summary: Get current authenticated user
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags: [auth]
      summary: Logout — revoke tokens and clear cookies
      description: Revokes the refresh token from the store and clears both auth cookies.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: logged out
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate refresh token and issue a new access token
      description: |
        Reads the `ota_refresh` cookie, validates it against the DB, deletes the old
        token (rotation), and issues a fresh access + refresh token pair.
      security: []
      responses:
        '200':
          description: Token refreshed; new cookies set
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: token refreshed
        '401':
          description: Missing or invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/delete-account:
    delete:
      tags: [auth]
      summary: Delete the authenticated user's account
      description: |
        Permanently deletes the account. Blocked if there are pending withdrawal requests.
        Revokes all refresh tokens and clears auth cookies.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Account deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: account deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Cannot delete — pending withdrawal exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Context / Topics ────────────────────────────────────────────────────────
...
security:
  - cookieAuth: []
  - bearerAuth: []